requirement.txt
selenium>='3.141.0'
mysql.connector>='2.2.9'
pandas>='1.0.1'